Added for Delete

+++ src/main/java/com/stmproject/repository/STM_HistoryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.stmproject.model.STM_Histo;
 
-public interface STM_HistoryRepository extends JpaRepository<STM_Histo, Long> {
-    
+public interface STM_HistoryRepository extends JpaRepository<STM_Histo, Integer> {
+
 }
